Fully separate symtab and symlite
The cmake LIGHTWEIGHT_SYMTAB option is documented to toggle between
symtab and symlite for ParseAPI, ProcControl, and Stackwalker.  However,
these libraries didn't fully commit one way or the other, so we were
left in limbo with both loaded most of the time.  This patch fixes them
to be fully separated into just symtab or just symlite.

- All CMakeLists now respect LIGHTWEIGHT_SYMTAB when setting target
  libraries, so they'll fail to link if something is wrong.

- ParseAPI now builds with symlite, but note that DyninstAPI won't work
  yet in this mode, because it assumes full SymtabCodeSource.

- ProcControl's getElfReader() now works with either symtab or symlite.

- Stackwalker's DebugStepperImpl::GetReg is refactored a bit to work
  without symtab, and the latent cap_stackwalker_use_symtab is now
  converted to WITH_SYMTAB_API and enabled.