symtabAPI: Add indirection between dwarf offsets and type IDs
The identity mapping between dwarf offsets and int type IDs is
problematic because offsets are actually 64-bit, so they may overflow an
int on 2GB+ files.  Offsets are also relative to their section, which
can be either .debug_info or .debug_types, so that overlap also needs to
be avoided.

The new DwarfWalker::type_ids_ maps between a pair offset,is_info to
typeId_t values incremented on demand.