symtabAPI: Automate context push/pop in DwarfWalker
In DwarfWalker::parse_int, there are many exit points between context push
and pop, especially for errors.  If the pop is missed, then the whole stack
may get into a bad state.  This patch adds a ContextGuard which manages the
push and pop via object lifetime, so all exit paths will be covered.