symtabAPI: Protect against GNU alt forms
Both DW_FORM_GNU_ref_alt and DW_FORM_GNU_strp_alt refer to an alternate
file, specified in .gnu_debugaltlink, which neither libdwarf or symtabAPI
currently understand.  We can at least recognize and skip past them though.

If we encounter a dwarf_diename() with an unhandled form, we now treat that
as if there weren't a name at all.  If we encounter a reference that's not
a normal offset, as with ref_alt, then this is now treated as a quiet
error, noted only in dwarf_printf.