proccontrol: Don't reference thread states too early
There was a chicken-and-egg in the int_thread constructor:

- int_thread calls the constructor for its first state.
- StateTracker calls its own setState.
- setState reads the up_thr handler and generator states,
  which aren't initialized yet!

Thankfully it didn't actually use those uninitized states unless
up_thr->up_thread is set, which isn't until after the initializer
list is done.  So it was harmless, but technically undefined behavior.

The fix is to not even read the generator and handler states until
up_thread is set.