parseAPI: return size_t from Function::num_blocks()
Changing to size_t from int better matches the size of the underlying type,
and also allows comparison to other sizes without any "signed vs. unsigned"
warnings.  This function is new to 8.2, so API/ABI isn't a concern.

A few places which used num_blocks() in dyninstAPI are also updated to
size_t, but none of these are in public headers.