cmake: Properly extend BUG_DEFINES on ppc linux
Both ppc32_linux and ppc64_linux set -Dbug_registers_after_exit in
BUG_DEFINES, but in doing so they clobbered the general linux bugs.  In
particular, -Dbug_syscall_changepc_rewind is still needed.  This patch
lets ppc just add its definition while keeping the others.

Our RHEL7 QA noticed this as a regression from 8.1.2, so I believe it
just broke as part of the transition to cmake.  In their smoke test, a
mutatee process in sleep() is attached, and it promptly crashes SIGILL
on the first iRPC.  I found that the iRPC was at 0x3fffb1f70000, and
proccontrol is setting that pc, but the crash is at 0x3fffb1f6fffc.  So
that appears to be the rewind issue, and a full BUG_DEFINES fixes it.

Reported-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Josh Stone <jistone@redhat.com>