dyninstAPI: fix EmitterIA32::emitDivImm for powers of 2
This tries to be clever and convert power-of-2 division into shifting
right, but the ia32 version had two bugs.  The call to emitMovRegToReg
should have had dest, then src, but this case was reversed.  Then the
raw opcodes for the intended SAR were accidentally encoding SAL.

The dest/src bug was reported by Coverity as SWAPPED_ARGUMENTS.