dyninstAPI: fix a few c_str() pointers outliving their string
A c_str() pointer is only good as long as its string lives, but a few
areas were using this on string return values that are then destructed.
For localized use, these now make sure to save the string locally.  Then
BPatch_variableExpr kept a const char* name, which now needs to be an
owned string member.

Reported by Coverity as WRAPPER_ESCAPE.