Adding a temporary program to test the functionality of the decoder table.
This program takes an instruction (in hex) as input and outputs the mnemonic of the corresponding instruction. Note that it does not check all corner cases, and hence may SEGFAULT if some entry in the decoder table cannot be found.

geninput.sh generates a file with instructions in hex on each line.

This program will be deleted later once testing is complete.