Added logic to re-order the operands after delayed decoding.
This is a temporary solution - will be replaced by a more generic solution in the coming weeks. For now, except for some instructions, the operand list is reversed after all operands are parsed. For the instructions that do not fit into this model, an operand swapping and rotating logic is used (at most there will be 1 swap and 2 rotates) to get the correct order.