Various DataflowAPI, DyninstAPI, PatchAPI, and runtime fixes.
+Stack analysis:
Minor changes to our definitions of transfer functions to produce more
consistent results when calculating stack heights.

+Dynamic target expressions:
This expression works again on Windows and Linux.

+Springboard generation:
Fixes a null pointer bug that occurs when a trap is used instead of a jump
to reach relocated code.

+Windows traps:
Using traps for instrumentation is now enabled on Windows builds.

+Error propagation:
Errors applying instrumentation in PatchAPI now propagate to the
DyninstAPI layer.

+Runtime variables:
Restores a missing runtime variable.