Error propagation and trusted code in defensive mode.
+Error checking and propagation:
Defensive mode parses all modules at start-up and instruments most
functions. This commit checks for errors and propagates them to the
upper layer. If any instrumentation fails, an error message is printed
and the process is aborted.

+Trusted code:
Defensive mode now uses a whitelist to mark certain system libraries as
safe. These are not parsed or instrumented at start-up. A new mechanism
exists for switching safe modules to defensive at runtime.