Truncate PTRACE_GETEVENTMSG exit status to int
The sys/wait.h macros are expecting to operate on an int, and they have
some ugly pointer macros to deal with old BSD compatibility.  But we get
an unsigned long from PTRACE_GETEVENTMSG.  Normally this work out ok,
but on big-endian ppc64 those macros end up reading the int from the
most significant bits only, which are zero.

Cast the long down to a local int first, so WEXITSTATUS works properly.

Fixes #35.