Semantics for FMOV
Added semantics manaully for FMOV. Floating point operations ar enot yet supported; however, the variant of FMOV we are seeing in jump tables does nothing but move a value from a FP register to a general purpose register. To allow the analysis to continue, the semantics for this instruction are thus manually added for now.