Modified implementation of DispatcherARM64::ConditionHolds()
The previous implementation did not make use of the fact the condition value in the expression can be retrieved as a raw integer value and does not have to be treated as a SValuePtr. Treating as an integer greatly simplifies assignment expression for all flags and also makes the code more readable.