Class declaration for base architecture specific formatter
The new strategy for instruction formatting is to have a class for each
architecture which inherit from a base class. All methods of the latter
are to be pure virtual, each representing formatting of a different type
of leaf node that can appear in the AST (register, immediate, dereference) and a method for re-ordering the operands if necessary that also returning the final instruction string.