Modifications to use the new ArchSpecificFormatter class for instruction/operand formatting
The signature of the format() method in InstructionAST and all inherited
classes that override it is modified to include a pointer to an
ArchSpecificFormatter object. The methods for each operand type
(Register, Immediate and Dereference) call the appropriate method on
ArchSpecificFormatter. The pointer is initialized within the constructor
of the Instruction class which also includes a new method to retrieve it
from elsewhere (getFormatter()).

This commit also includes implementation of methods of the ArmFormatter
class. Instruction::format() currently returns an empty string for architectures other
than ARM64. Implementation for other architectures will be added in
future commits.