Semantics for UBFM and SBFM variants
This commit introduces semantics for the 5 UBFM variants (UXTB, UXTH, UBFM, UBFIZ, UBFX) and SBFM variants (SXTB, SXTH, SBFM, SBFIZ, SBFX), and a couple functions in the DispatcherARM64 class used by these semantics. Implementation of one of these functions needs to be updated (next commit) and UXTB's semantics are modified to proceed with jump table analysis for now; the original semantics are still present as comments and will be restored in a future commit.