Conservative fix for the stack pointer passing problem.
If stack pointers may be passed as parameters to a callee, avoid
instrumenting the caller.  Doing so may result in the pointers not
being adjusted correctly for usage in the callee.  Check pointers
in all registers designated as parameter registers in the ABI, as
well as the last 7 stack slots.