Create classes for Definition-Height pairs and sets of those pairs.
The DefHeight typedef is now a class, thereby allowing us to access
its members by name rather than "first" and "second". In addition
there is now a wrapper class for sets of DefHeights with several
useful methods built in for operations on such sets. Previously we
were performing those operations with static methods in the
StackAnalysis class.