dyninst parses some function symbol as knownData
If symbol A and B share the same Region and Offset, A's type is
FUNCTION, however B is not, it will casue bug that the B's address
will be added into @KnownData. Here we double check whether there
existing another symbol who shares the same Region and Offset with
the current, but its type is FUNCTION, we should prevent the symbol's
address from being added into @knownData.

symbol "__aesni_set_encrypt_key"(type: NOTYPE) and
"aesni_set_encrypt"(type: FUNCTION) in libcrypto.so.1.1 meet
the condition.