Fixes for dynamic function wrapping without PLT entries
Fixes dynamic function wrapping to modify the GOT directly in cases
where a PLT stub does not exist. Previously, dyninst would search
PLT entries for the symbol for which the original function being wrapped
would be relocated to. If not found, the GOT entry for that symbol would
not be modified leading to undesired results. GCC can create code
that directly looks a function up within the GOT inline, any function that
was attempted to be wrapped that had this optimization would silently fail.