Mimic PLT behavior when replacing a call to a PLT stub [PPC]
PLT functions on Power have an instruction that stores the current TOC (R2)
onto the stack at position 24 (i.e. std r2,24(r1)). A new stack frame is not
created by the PLT, R2 is stored in the callers frame.

Upon return, the function that called the PLT stub tries to restore its
TOC from the stack (i.e. "bl pltStub" followed by a "ld r2,24(r1)").
We need to ensure that R2 is on the callers stack for any PLT function
we replace.

What this patch does is for any call to a PLT stub that has been modified,
we insert "std r2,24(r1)" to simulate that a PLT was called (even if we are
calling a function direct). The new pattern written is the following:

if original call was PLT:
	write std r2,24(r1)
<Original CFWidget Call Generation for PPC>

We determine if the original call was a PLT by first saving the original target
of the call instruction in Modification.C, passing this address to the widget,
then checking if the original target address is listed in the image's pltstub list.

 expects to be able to restore R2 from stack position 24