Improve detection of ghost threads (#784)
Change plat_handle_ghost_thread to return 'bool' (see #742).

This indicates if the thread is actually a ghost or not. This can happen
because ptrace returned ESRCH (no such process) in
LinuxPtrace::ptrace_int, yet the process is alive. This is a valid state
for ptrace, but ptrace_int doesn't check for this. The ptrace man-page
indicates that this can happen when a process received a STOP signal,
but it hasn't transitioned to the new state yet.