Cleanup orphaned code (#1064)
* Remove 'dummy.C' from dyninstAPI/CMakeLists.txt

* Remove unused scripts/buildstamp

* Remove unused files from parseAPI/CMakeLists.txt

* Remove instructionAPI/src/MainDocPage.C

The usage of this file was removed by 5c88375a7 in 2016.

* Remove stackwalk/src/firstparty.C

This file was abandoned by 44507e6b0 in 2013.

* Remove BPatch_dependenceGraph{Edge,Node}.{C,h}

These files were abandoned by bbc9ff655 in 2009.

* Remove dyninstAPI/src/dummy.C

This file was abandoned by 44507e6b0 in 2013.

* Remove dyninstAPI/src/Relocation/Widgets/CFPatch.C

This has been abandoned since at least 91246b9ca in 2011.

* Remove dyninstAPI/src/InstructionCache.C

This file was never compiled. It seems to be an outdated copy of dataflowAPI/src/InstructionCache.C.

* Remove dyninstAPI/src/BPatch_templates.C

This was orphaned by 44507e6b0 in 2013.

* Remove symtabAPI/src/staticRewrite.C

This was orphaned by 44507e6b0 in 2013.

* Remove dyninstAPI_RT/src/RTheap-svr4.c

This was orphaned by 44507e6b0 in 2013.

* Remove dyninstAPI_RT/src/RTmutatedBinary{_ELF}.c

These were orphaned sometime before 44507e6b0 in 2013. The git history doesn't show when. I'm assuming it was lost in the transition from CVS.

* Remove dyninstAPI_RT/src/RTthread-power.c

This was a holdover from the rs6000 AIX support.

* Remove common/src/tunableConst.h

This has never been part of the build system.

* Remove common/src/List.{h,C}

These were orphaned by 44507e6b0 in 2013.

* Remove common/src/auxvtypes.{h,C}

These should have been removed by cccad8ee.

* Remove common/src/int64iostream.{h,C}

These were orphaned by 44507e6b0 in 2013.

* Remove unused 'timing.h' interface

* Remove unused dyninstAPI/src/clock.h

* Remove unused common/src/mmtimer.c

* Remove common/src/debugOstream.{h,C}

This was abandoned by by 65bd8055 in 2008

* Remove unused common/src/Time.{h,C}

* Remove unused common/src/fraction.{h,C}

* Remove symtabAPI/src/Symtab-deprecated.C

There are no symbols defined by this file.

* Remove common/src/Ident.{h,C}

This was orphaned by 44507e6b0 in 2013.

* Remove parseThat/src/dyninstCompat.v4.C

This should have been removed as part of 63d791419.

* Remove common/src/addrtranslate-static.C

This was orphaned by 44507e6b0 in 2013.

* Remove dyninstAPI/src/asmExterns.h

This was abandoned by 5019a9c5a in 2007.

* Remove dyninstAPI/src/BPatch_typePrivate.h

This was abandoned by b3105c49c in 2008.

* Remove dyninstAPI/src/codebuffer.h

This was added by 936238d in 2010, but never added to the build system.

* Remove parseAPI/h/GraphAdapter.h

This was abandoned by 44507e6b0 in 2013.

* Remove common/src/cstring.h

This was abandoned by f4613dcfe in 2008.

* Remove common/src/keylist.h

This was abandoned by f4613dcfe in 2008.

* Remove common/src/klist.h

This was abandoned by f4613dcfe in 2008.

* Remove common/src/Line.h

This was abandoned by 4b1956db in 2013.

* Remove stackwalk/h/local_var.h

This was abandoned by 44507e6b0 in 2013.

* Remove common/src/LockTraits.h

Was added by a38bacc5 in 2017, but never added to the build system.

* Remove common/src/matherr.h

This was added by 4b1956db in 2013, but was never part of the build system.

* Remove dyninstAPI/src/ntDL.h

This was abandoned by 5c88375a in 2016.

* Remove dyninstAPI/src/NTTypes.h

This was abandoned by 5c88375a78 in 2016.

* Remove dataflowAPI/src/Operations.h

This was abandoned by 5c88375a in 2016.

* Remove common/h/stl_utils.h

This should have been removed by cccad8ee.

* Remove common/src/string-regex.{h,C}

This was abandoned by f1b11229 in 2008.

* Remove stackwalk/{src,h}/sw_c.{h,C}

This was abanded by 44507e6 in 2013.

* Remove elf/src/SystemTap.{h,C}

This was abandoned by 44507e6 in 2013.

* Remove common/src/triple.h

This was abandoned by 5c88375a7 in 2016.

* Remove dataflowAPI/rose/x86InstructionSemantics-notemplate.h

This was added by 8d6626c in 2010, but was never part of the build system.

* Remove unused man files from common/docs

* Remove declaration of 'friend operator<<' from BPatch_basicBlockLoop.h

Its implementation was removed by f3b69f6c8 in 2005.

* Remove declaration of 'friend operator<<' from BPatch_flowGraph.h

Its implementation was removed by f3b69f6c8 in 2005.

* Remove declaration of 'friend operator<<' from BPatch_frame.h

It introduced by 954046f99 in 2005, but never implemented.

* Remove declaration of 'friend operator<<' from CFG.h/Loop

It introduced by 6b2b85ee5 in 2014, but never implemented.

* Remove 'operator<<' from BPatch_sourceBlock

This was abandoned by 0d9f1455cc in 2003.